Move TranslationsAsI18n to Translations.ts to fix circular dependency
