+++ packages/core/src/index.ts
@@ -124,7 +124,7 @@ export { default as StencilShapeRegistry } from './view/geometry/node/StencilSha
 export * as constants from './util/Constants';
 export { default as Guide } from './view/other/Guide';
 
-export { default as Translations } from './i18n/Translations';
+export { default as Translations, TranslationsAsI18n } from './i18n/Translations';
 export * from './i18n/config';
 export * from './i18n/provider';
